Split policy-related models into modules
